(refactor) move diffusion utility functions to separate namespace class
